+++ app/Http/Controllers/Api/AuthController.php
@@ -3,13 +3,12 @@
 namespace App\Http\Controllers\Api;
 
 use App\Exceptions\User\LoginFailedException;
+use App\Http\Requests\User\RegisterUserRequest;
 use App\Http\Resources\UserResource;
-use App\Models\User;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\User\LoginUserRequest;
-use App\Http\Requests\RegisterUser;
 use App\UseCases\User\LoginUser;
-use App\ViewModels\UserViewModel;
+use App\UseCases\User\RegisterUser;
 use Illuminate\Support\Facades\Auth;
 
 class AuthController extends Controller
@@ -29,10 +28,8 @@ class AuthController extends Controller
         return new UserResource(Auth::user());
     }
 
-    public function register(RegisterUser $request)
+    public function register(RegisterUserRequest $request, RegisterUser $usecase)
     {
-        $user = User::create($request->toAttributes());
-
-        return new UserViewModel($user);
+        return new UserResource($usecase($request->makeUser()));
     }
 }
